optimize CUDA hot path: FP32-only textures, divergence-free RNG/ONB, warp-aligned blocks
Targets the per-bounce shading loop and warp utilization without any
architectural rewrites.

P0 — correctness/large perf hazards:
- Texture::value: double->float (Texture.h). Eliminates silent FP64 path
  on every shading hit (CheckerTexture used double sin() and the virtual
  base forced double conversion at every call site). ImageTexture uses
  __saturatef instead of fminf(fmaxf(...,0),1).
- random_in_unit_sphere: replace rejection loop with closed-form
  spherical sampling (Material.h). Original loop diverged until the last
  warp lane accepted (~5-6 iterations vs 1.91 mean per lane). cbrtf
  radial scale preserves volume-uniform distribution for behavioral
  parity with Metal::scatter / SubsurfaceScatter / random_in_hemisphere.
- __sincosf for paired cos/sin in cosine_weighted_hemisphere
  (Material.h), Sphere::sample_point, Disc::sample_point.

P1 — tractable wins:
- Block geometry: 16x16 -> 32x8 in raytrace launches (kernel.cu render,
  render_to_buffer, render_tile). Warp now covers one 32-pixel row,
  coalescing framebuffer writes and row-aligning per-lane Philox state
  reads.
- Replace `nx / tx + 1` with `(nx + tx - 1) / tx` in 5 launch sites to
  avoid an extra empty block when nx % tx == 0.
- Drop the 3 cudaDeviceSynchronize calls after render_init —
  stream-ordering already serializes the next kernel against the rand
  state init.
- Branchless ONB (Frisvad-Duff revised) in build_onb (Material.h),
  Disc constructor, and TriangleMesh hit-record fallback paths.
  Removes the divergent fabsf(n.x) > 0.9 branch.

P2 — FP64 contamination cleanup:
- World::surrounding_box: fmin/fmax -> fminf/fmaxf.
- Sphere::hit: sqrt -> sqrtf (also CSE the sqrt across both roots).
- Material.h: 1e-8 -> 1e-8f, fabs -> fabsf, 1.0 -> 1.0f, and `> 0.0`
  hemisphere check -> `> 0.0f`.

Deferred (need new files / golden-image verification):
- P0-3 flat top-level BVH (replaces virtual recursive BVHNode::hit).
- P0-4 cached AABB lookups in BVH builder.
- P1-2 curand_uniform4 batching.
- P1-4 final/tagged-dispatch for Material/Hittable.
- P1-5 lights/BVH in __constant__ or shared.
- P1-7 cudaTextureObject_t for ImageTexture.

+++ src/raytracer/World.h
@@ -118,13 +118,13 @@ __device__ World::~World() {
 }
 
 __device__ AABB surrounding_box(AABB box0, AABB box1) {
-    glm::vec3 lo(fmin(box0.min().x, box1.min().x),
-        fmin(box0.min().y, box1.min().y),
-        fmin(box0.min().z, box1.min().z));
+    glm::vec3 lo(fminf(box0.min().x, box1.min().x),
+        fminf(box0.min().y, box1.min().y),
+        fminf(box0.min().z, box1.min().z));
 
-    glm::vec3 hi(fmax(box0.max().x, box1.max().x),
-        fmax(box0.max().y, box1.max().y),
-        fmax(box0.max().z, box1.max().z));
+    glm::vec3 hi(fmaxf(box0.max().x, box1.max().x),
+        fmaxf(box0.max().y, box1.max().y),
+        fmaxf(box0.max().z, box1.max().z));
 
     return AABB(lo, hi);
 }
